Move logics from a base class to trait
